- Fix updater
- Add support for open repo menu
- Add support for commit from within diff
- Mark long diff lines with a '...' char
- Add support for undo/restore all binary files in diff
- Add support in diff for main menu with scroll to and undo
- Refresh diff after undo
- Add support for undo/restore files in diff
- Adjust push branch and enhanse push/fetch logging
- Add support for refresh diff of uncommitted changes
- Refactor diff copy
- Fix issue with ' ' in file paths when diff and status/undo
- Try to fix path with space in clone
- Add install file exists check
- Adjust check mark signs
- Pause file monitor during git commit
- Trying different folder montitor event handling

+++ gmd/Installation/Updater.cs
@@ -32,7 +32,6 @@ public class GitAsset
     public string browser_download_url { get; set; } = "";
 }
 
-
 class Updater : IUpdater
 {
     static readonly TimeSpan checkUpdateInterval = TimeSpan.FromHours(1);
@@ -200,15 +199,21 @@ class Updater : IUpdater
         {
             // Move downloaded file next to this process file (replace must be on same volume)
             var newPath = GetTempPath();
+            if (!Files.Exists(downloadedPath))
+            {
+                Log.Info($"No new file to install {downloadedPath}");
+                return R.Ok; // No new file to install, some other thread already installed it
+            }
+
             File.Move(downloadedPath, newPath);
-            Log.Info($"Install {newPath} ...");
+            Log.Info($"Move {downloadedPath} => {newPath} ...");
             if (!Try(out var e, MakeBinaryExecutable(newPath))) return e;
 
             var thisPath = Environment.ProcessPath ?? "gmd";
             var newThisPath = GetTempPath();
 
             File.Move(thisPath, newThisPath);
-            Log.Info($"Moved {thisPath} to {newThisPath}");
+            Log.Info($"Moved {thisPath} => {newThisPath}");
             File.Move(newPath, thisPath);
             Log.Info($"Installed {newPath}");
             return R.Ok;
@@ -483,7 +488,7 @@ class Updater : IUpdater
             return R.Ok;
         }
 
-        return cmd.Run("chmod", $"+x {path}", "");
+        return cmd.Command("chmod", $"+x {path}", "");
     }
 
     private bool IsDotNet()
